refactor: remove the migrate command and the interactive TUI surface
Two dead command surfaces, ~4.2k lines net.

1. `aimee migrate v2` could not work.

   It routed to migrate.v2, which the server has no handler for — and
   test_server_dispatch.c ASSERTS that it answers UNKNOWN_METHOD, so the
   server side was deliberately retired and pinned that way. The client still
   printed "usage: aimee migrate v2 <legacy-store>" and marshalled a request
   that was always rejected. build-integrity independently lists "migrate" as a
   storage-internal term that must never appear in client help, which is the
   same fact from the other side.

   Removes the route row, marshal_storage_migrate, pt_print_migrate_v2 +
   print_storage_migrate + the printer registration, and the header declaration
   that was duplicated twice. The UNKNOWN_METHOD assertion stays as the guard
   that it does not come back.

2. The interactive TUI is gone, and bare `aimee` prints usage.

   `aimee` with no arguments forwarded a "launch" to the server and exec'd a
   provider TUI — the bare command did the single heaviest thing available, and
   off a TTY it exited 0 having printed nothing at all. It now prints usage.
   `aimee chat`, the OpenCode frontend (cli_tui_opencode_v2.c + v2b.c) and the
   native fallback loop go with it.

   What survives is NOT a TUI: cli_chat_stream() is the headless /v1 streaming
   path, and acp-serve is its only caller (editors like Zed). It lived in
   cli_tui.c, so cli_tui.{c,h} is renamed to cli_chat_stream.{c,h} — 982 lines
   down to ~400, none of which touch a terminal. The name was the only reason
   to think this code was interactive.

   The dead set was found by the compiler rather than by eye: deleting the entry
   points and rebuilding under -Werror surfaced each newly-unused static in turn
   (builtin_chat_send, chat_v1_post, chat_v1_session_path, the presence
   attach/detach pair, the codex/claude config model-detection helpers).
   builtin_chat_send_streaming_control had NO callers at all and survived only
   because non-static functions raise no unused warning.

   Fallout handled:
   - CMakeLists.txt still listed the deleted TUs; windows-cmake would have failed.
   - test_cli.sh asserted `aimee chat --help` mentions OpenCode, and its
     "opencode v2 bridge contract" case invoked bare `aimee` — i.e. it tested the
     exact behavior being removed. Replaced with a bare-invocation usage check;
     opencode_v2_probe.sh (310 lines) went with it.
   - ~20 build-integrity assertions pinned the TUI as WIRED (prompt extraction,
     queued-turn acks, GlobalEvent envelopes, aggregateID sync). They pinned code
     that no longer exists, so they are replaced by the inverse: one guard that
     the retired frontends stay absent, extending the existing Codex one. It
     immediately earned its keep by catching the stale CMake entries and two of
     my own comments naming the retired symbols.
   - AIMEE_MODEL / AIMEE_EFFORT / AIMEE_OPENCODE_BIN were read only by the
     deleted model-detection helpers; the env-var reference drops 155 -> 152.

Verified: make all, make lint, make unit-tests ("All tests passed.") and
make build-integrity all exit 0 locally, with the new guard passing.
