+++ src/headers/cli_v1_routes_internal.h
@@ -170,7 +170,6 @@ void pt_print_memory_read(const char *method, cJSON *resp);
 void pt_print_memory_search(const char *method, cJSON *resp);
 void pt_print_memory_stats(const char *method, cJSON *resp);
 void pt_print_memory_store(const char *method, cJSON *resp);
-void pt_print_migrate_v2(const char *method, cJSON *resp);
 void pt_print_model_list(const char *method, cJSON *resp);
 void pt_print_model_refresh(const char *method, cJSON *resp);
 void pt_print_model_show(const char *method, cJSON *resp);
@@ -386,7 +385,6 @@ void pt_print_memory_read(const char *method, cJSON *resp);
 void pt_print_memory_search(const char *method, cJSON *resp);
 void pt_print_memory_stats(const char *method, cJSON *resp);
 void pt_print_memory_store(const char *method, cJSON *resp);
-void pt_print_migrate_v2(const char *method, cJSON *resp);
 void pt_print_model_list(const char *method, cJSON *resp);
 void pt_print_model_refresh(const char *method, cJSON *resp);
 void pt_print_model_show(const char *method, cJSON *resp);
